requirements.txt
# =============================================================================
# Capstone Project — requirements.txt
# Python: 3.11 (khuyến nghị 3.11.9)
# =============================================================================
# TRƯỚC KHI CHẠY FILE NÀY, cài thủ công theo thứ tự:
#
# 1. pip install numpy==1.26.4
#
# 2. Cài PyTorch theo nền tảng:
#    CUDA 12.1:  pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#    CPU only:   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#    macOS MPS:  pip install torch torchvision
#
# 3. Cài torchreid từ GitHub:
#    pip install git+https://github.com/KaiyangZhou/deep-person-reid.git
#    (Nếu lỗi trên Windows: clone thủ công rồi pip install -e .)
#
# 4. pip install -r requirements.txt
# =============================================================================


# -----------------------------------------------------------------------------
# [1] COMPUTER VISION — Detection & Pose
# -----------------------------------------------------------------------------
ultralytics==8.3.55
opencv-python==4.10.0.84
pycocotools==2.0.8
# opencv-python-headless==4.10.0.84  # dùng thay thế nếu server/embedded không có display


# -----------------------------------------------------------------------------
# [2] FACE RECOGNITION
# -----------------------------------------------------------------------------
insightface==0.7.3
onnxruntime==1.19.2
# onnxruntime-gpu==1.19.2  # thay thế nếu có GPU CUDA (không dùng cả hai)


# -----------------------------------------------------------------------------
# [3] VECTOR DATABASE
# -----------------------------------------------------------------------------
faiss-cpu==1.9.0
# faiss-gpu==1.7.4  # thay thế nếu CUDA, chỉ Linux


# -----------------------------------------------------------------------------
# [4] WEB FRAMEWORK — Flask
# -----------------------------------------------------------------------------
flask==3.1.0
flask-socketio==5.4.1
flask-cors==5.0.0
python-socketio==5.11.4
eventlet==0.37.0
# gevent==24.2.1  # thay thế eventlet nếu lỗi trên macOS ARM


# -----------------------------------------------------------------------------
# [5] WEB FRAMEWORK — FastAPI
# -----------------------------------------------------------------------------
fastapi[standard]==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
httpx==0.28.1
pydantic==2.10.3


# -----------------------------------------------------------------------------
# [6] DATA & NUMERICAL
# -----------------------------------------------------------------------------
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3


# -----------------------------------------------------------------------------
# [7] VIDEO I/O & IMAGE
# -----------------------------------------------------------------------------
av==13.1.0
imageio==2.36.1
Pillow==11.0.0


# -----------------------------------------------------------------------------
# [8] CONFIGURATION & UTILITIES
# -----------------------------------------------------------------------------
pyyaml==6.0.2
python-dotenv==1.0.1
rich==13.9.4
click==8.1.8
tqdm==4.67.1
omegaconf==2.3.0
gdown==6.0.0


# -----------------------------------------------------------------------------
# [9] TRACKING
# -----------------------------------------------------------------------------
filterpy==1.4.5
lapjv
# lap==0.5.12  # thay thế lapjv nếu đã có MSVC build tools trên Windows


# -----------------------------------------------------------------------------
# [10] LOGGING
# -----------------------------------------------------------------------------
loguru==0.7.3


# -----------------------------------------------------------------------------
# [11] DEEP SORT (body tracking alternative)
# -----------------------------------------------------------------------------
deep-sort-realtime==1.3.2


# -----------------------------------------------------------------------------
# [12] ONNX EXPORT & INFERENCE
# -----------------------------------------------------------------------------
onnx==1.21.0


# -----------------------------------------------------------------------------
# [13] UTILITIES
# -----------------------------------------------------------------------------
requests==2.33.1
psutil==7.2.2
packaging==26.1
python-dateutil==2.9.0.post0
six==1.17.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
urllib3==2.6.3
sympy==1.14.0
networkx==3.6.1
filelock==3.29.0
fsspec==2026.3.0
typing_extensions==4.15.0


# =============================================================================
# requirements-dev.txt — KHÔNG cài trên production/embedded
# Tạo file riêng: requirements-dev.txt rồi: pip install -r requirements-dev.txt
# =============================================================================
# pytest==8.3.4
# pytest-asyncio==0.24.0
# tensorboard==2.20.0
# matplotlib==3.10.8
# scikit-learn==1.8.0
# scikit-image==0.26.0
# ipykernel==7.2.0
# ipython==9.12.0
# jupyter_client==8.8.0
# jupyter_core==5.9.1
# polars==1.40.0
# prettytable==3.17.0
# albumentations==2.0.8
